# Core code refinements